Ensure dfmem erases last sector before proceeding
By construction, dfmem functions don't block execution while the memory
is processing a command until they need to issue the next command to it.
This causes the sector erase loop to quit and start the sensor capture
before the memory is ready to save any data. After three rows are
queued, the dfmem function that saves the buffer to memory has to busy
wait until the erase finishes, introducing a large delay before the
4th sample. The commit should fix this problem.